requirements.txt
requests
python-dotenv
google-genai
statistics; python_version < '3.8'
